requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Authlib
python-dotenv
Flask-Login
Flask-WTF
email-validator